// src/common/enums/story-type.enum.ts
export enum StoryType {
  FANTASY = 'FANTASY',
  ADVENTURE = 'ADVENTURE',
  SCI_FI = 'SCI_FI',
  MYSTERY = 'MYSTERY',
  HORROR = 'HORROR',
  ROMANCE = 'ROMANCE',
  COMEDY = 'COMEDY',
  DRAMA = 'DRAMA',
  HISTORICAL = 'HISTORICAL',
  FAIRY_TALE = 'FAIRY_TALE',
  CHILDREN = 'CHILDREN',
  ACTION = 'ACTION',
  THRILLER = 'THRILLER',
  LORD_OF_THE_RINGS = 'LORD_OF_THE_RINGS',
  MIDDLE_EARTH_FANTASY = 'MIDDLE_EARTH_FANTASY',
  HOBBIT_FANTASY = 'HOBBIT_FANTASY',
  EPIC_FANTASY = 'EPIC_FANTASY',
  HIGH_FANTASY = 'HIGH_FANTASY',
  DARK_FANTASY = 'DARK_FANTASY',
}